feat: 1.完成顯示好友列表功能 2.新增貼文按讚及收回功能、顯示貼文總讚數功能

+++ server/controllers/users.js
@@ -6,6 +6,7 @@ export const getUser = async (req, res) => {
   try {
     const { id } = req.params;
     const user = await User.findOne({ where: { id: id } });
+
     res.status(200).json(user);
   } catch (err) {
     res.status(500).json(err);
@@ -31,11 +32,13 @@ export const addRemoveFriend = async (req, res) => {
     const { id, friendId } = req.params;
     const user = await Relationship.findAll({ where: { followerUserId: id }, raw: true, attributes: ["followedUserId"] });
     const formattedFollowedUserId = await Promise.all(user.map((x) => x.followedUserId));
+
     if (formattedFollowedUserId.includes(+friendId)) {
-      await Relationship.destroy({ where: { followerUserId: +id } });
+      await Relationship.destroy({ where: { followerUserId: +id, followedUserId: +friendId } });
     } else {
       await Relationship.create({ followerUserId: +id, followedUserId: +friendId });
     }
+
     res.status(200).json(`Friendlist updated`);
   } catch (err) {
     res.status(500).json(err);
